Missing classname
git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@2761 e03df62e-2008-0410-955e-edbf42e46eb7

+++ src/inspircd.cpp
@@ -289,7 +289,7 @@ void InspIRCd::MoveTo(std::string modulename,int slot)
 			break;
 		}
 	}
-	if (v2 == slot)
+	if (v2 == (unsigned int)slot)
 	{
 		log(DEBUG,"Item %s already in slot %d!",modulename.c_str(),slot);
 	}
